Add AbstractDbTimestampBoundStore
